chore: extract `selection` module out of `modeling`
