--- src/trackingModules/diagram/index.js
@@ -1,7 +1,0 @@
-import selectionTracking from './selection';
-
-export default {
-  __depends__: [
-    selectionTracking
-  ]
-};
